Tweak CSS for tables; add support for 3-...-15 cats.
The CSS for tables now alternatively colors the rows in the
category list using cyan and blue.  Future work may involve
having two sets of alternations: one for days and another
for categories within a day.

Support has also been added to the write_category.php form
to allow writers to specify the point value scheme to be
applied to the category.  In addition to the 2-4-6-8-10
scheme which was used by default, there is now the option
for a 3-6-9-12-15 scheme also.