Add support for stripping tags from question data.
This change strips all tags from categories, explantory texts
clues and responses when they are entered into the database.  The
<b>, <i>, and <u> tags are allowed in category titles, explanatory
text, and clues, and additionally, in explanatory text and
clues, the <a> tag is also allowed.  Other tags in those fields,
as well as all HTML tags in responses (which should be regular
expressions), are stripped before entering the database.

Additionally, category.php strips HTML tags from the category
displayed in the title element.