autofill: restructure code & response cleaning
Some of the code in the script has been split off into functions.

The response-cleaning code in autofill_cron is now in a common function,
starting from the point where the response text has been isolated from
the post.  Also, it has been modified to (finally!) not reject
unicode letters, so we won't be having any more embarrasing hiccups with
responses containing accented characters, diacritics, and the like.